Convert `SocketFactory` into a class, instead of a interface
